rename functions to be coherent, split behaviours into files
